Create simple endpoint for barcode decode
