+++ InvenTree/InvenTree/views.py
@@ -22,7 +22,6 @@ from common.models import InvenTreeSetting
 
 from .forms import DeleteForm, EditUserForm, SetPasswordForm
 from .helpers import str2bool
-from .version import inventreeVersion, inventreeInstanceName
 
 from rest_framework import views
 
@@ -414,22 +413,6 @@ class AjaxDeleteView(AjaxMixin, UpdateView):
         }
 
         return self.renderJsonResponse(request, form, data=data, context=context)
-
-
-class InfoView(AjaxView):
-    """ Simple JSON endpoint for InvenTree information.
-    Use to confirm that the server is running, etc.
-    """
-
-    def get(self, request, *args, **kwargs):
-
-        data = {
-            'server': 'InvenTree',
-            'version': inventreeVersion(),
-            'instance': inventreeInstanceName(),
-        }
-
-        return JsonResponse(data)
 
 
 class EditUserView(AjaxUpdateView):
